pyproject.toml
[project]
name = "MECH-M-DUAL-2-MLB-DATA"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Peter Kandolf", email = "peter.kandolf@mci.edu"},
]
dependencies = ["numpy>=2.2.3", "scipy>=1.15.2", "requests>=2.32.3", "scikit-learn>=1.6.1", "skl2onnx>=1.18.0", "onnxruntime>=1.21.0", "skops>=0.11.0", "dvc>=3.59.1", "pathlib>=1.0.1", "logging>=0.4.9.6", "omegaconf>=2.3.0", "dvclive>=3.48.2", "dvc-webdav>=3.0.0", "pywavelets>=1.8.0", "gitpython>=3.1.44"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
